requirements.txt
requests
boto3
pyspark
google-cloud-bigquery
python-dotenv
apache-airflow